+++ test/prompt_interface/test_editor_vibe_visual.rb
@@ -64,6 +64,39 @@ class TestPromptInterfaceEditorVibeVisual < KwardTestCase
     end
   end
 
+  def test_prompt_interface_vibe_mode_visual_inserts_cursors_at_selected_line_boundaries
+    Dir.mktmpdir do |dir|
+      File.write(File.join(dir, "notes.txt"), "  one\ntwo longer\n three")
+      Dir.chdir(dir) do
+        prompt = Kward::PromptInterface.new(input: StringIO.new, output: StringIO.new, editor_mode: "vibe")
+        assert prompt.send(:open_editor, "notes.txt")
+        editor = prompt.instance_variable_get(:@editor_state)
+
+        prompt.send(:handle_editor_key, "v")
+        2.times { prompt.send(:handle_editor_key, "j") }
+        prompt.send(:handle_editor_key, "I")
+
+        assert_equal "insert", editor.vibe_mode
+        assert_equal [0, 6, 17], editor.selections.map { |selection| selection[:cursor] }
+        ">".each_char { |char| prompt.send(:handle_editor_key, char) }
+        prompt.send(:handle_editor_key, "\e")
+        assert_equal ">  one\n>two longer\n> three", editor.buffer
+
+        editor.buffer = "  one\ntwo longer\n three"
+        editor.cursor = 0
+        prompt.send(:handle_editor_key, "v")
+        2.times { prompt.send(:handle_editor_key, "j") }
+        prompt.send(:handle_editor_key, "A")
+
+        assert_equal "insert", editor.vibe_mode
+        assert_equal [5, 16, 23], editor.selections.map { |selection| selection[:cursor] }
+        "!".each_char { |char| prompt.send(:handle_editor_key, char) }
+        prompt.send(:handle_editor_key, "\e")
+        assert_equal "  one!\ntwo longer!\n three!", editor.buffer
+      end
+    end
+  end
+
   def test_prompt_interface_vibe_mode_visual_block_inserts_and_appends_text
     Dir.mktmpdir do |dir|
       File.write(File.join(dir, "notes.txt"), "one\ntwo\nthree")
@@ -79,7 +112,7 @@ class TestPromptInterfaceEditorVibeVisual < KwardTestCase
         prompt.send(:handle_editor_key, "\e")
         assert_equal "# one\n# two\n# three", editor.buffer
 
-        editor.buffer = "one\ntwo\nthree"
+        editor.buffer = "one\ntwenty\nthree"
         editor.cursor = 0
         editor.set_cursor_line_and_column(0, 2)
         prompt.send(:handle_editor_key, "\x16")
@@ -87,7 +120,7 @@ class TestPromptInterfaceEditorVibeVisual < KwardTestCase
         prompt.send(:handle_editor_key, "A")
         "!".each_char { |char| prompt.send(:handle_editor_key, char) }
         prompt.send(:handle_editor_key, "\e")
-        assert_equal "one!\ntwo!\nthree", editor.buffer
+        assert_equal "one!\ntwenty!\nthree", editor.buffer
       end
     end
   end